requirements.txt
pandas
pyspark
scikit-learn
boto3
apache-airflow
